+++ tests_basic/test_winstonlutz.py
@@ -49,10 +49,6 @@ from tests_basic.utils import (
 TEST_DIR = "Winston-Lutz"  # comment trigger
 
 
-def create_wl_dataset():
-    generate_winstonlutz()
-
-
 def apply_rotation_to_points(
     points: list[Point], angle: float | Sequence[float], axis: str
 ) -> list[Point]:
@@ -933,6 +929,7 @@ class WinstonLutzMixin(CloudFileMixin):
     cax2bb_max_distance = 0
     cax2bb_median_distance = 0
     cax2bb_mean_distance = 0
+    cax2epid_max_distance = 0
     epid_deviation = None
     bb_shift_vector = Vector()  # vector to place BB at iso
     machine_scale = MachineScale.IEC61217
@@ -1033,6 +1030,11 @@ class WinstonLutzMixin(CloudFileMixin):
             self.wl.cax2bb_distance(metric="median"),
             self.cax2bb_median_distance,
             delta=0.1,
+        )
+
+    def test_cax_to_epid_distance(self):
+        self.assertAlmostEqual(
+            self.wl.cax2epid_distance("max"), self.cax2epid_max_distance, delta=0.1
         )
 
     def test_bb_mean_distance(self):
@@ -1242,6 +1244,7 @@ class Synthetic1mmLeftNoCouch(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1
     cax2bb_mean_distance = 0.5
     cax2bb_median_distance = 0.5
+    cax2epid_max_distance = 1
 
 
 class Synthetic1mmLeft(SyntheticWLMixin, TestCase):
@@ -1251,6 +1254,7 @@ class Synthetic1mmLeft(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1
     cax2bb_mean_distance = 0.67
     cax2bb_median_distance = 1
+    cax2epid_max_distance = 1.02
     couch_iso_size = 2
 
 
@@ -1259,6 +1263,7 @@ class Synthetic1mmRight(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_mean_distance = 0.75
     cax2bb_median_distance = 1
+    cax2epid_max_distance = 1.02
     couch_iso_size = 2
 
 
@@ -1267,6 +1272,7 @@ class Synthetic1mmUp(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_mean_distance = 0.25
     cax2bb_median_distance = 0
+    cax2epid_max_distance = 1
 
 
 class Synthetic1mmDown(SyntheticWLMixin, TestCase):
@@ -1274,6 +1280,7 @@ class Synthetic1mmDown(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_mean_distance = 0.25
     cax2bb_median_distance = 0
+    cax2epid_max_distance = 1
     couch_iso_size = 0
 
 
@@ -1282,6 +1289,7 @@ class Synthetic1mmIn(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_mean_distance = 1.0
     cax2bb_median_distance = 1
+    cax2epid_max_distance = 1.02
     couch_iso_size = 2.0
 
 
@@ -1291,6 +1299,7 @@ class Synthetic1mmOut(SyntheticWLMixin, TestCase):
     cax2bb_mean_distance = 1.0
     cax2bb_median_distance = 1
     couch_iso_size = 2.0
+    cax2epid_max_distance = 1
 
 
 class Synthetic1mmIn1mmLeft(SyntheticWLMixin, TestCase):
@@ -1299,6 +1308,7 @@ class Synthetic1mmIn1mmLeft(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.41
     cax2bb_mean_distance = 1.3
     cax2bb_median_distance = 1.4
+    cax2epid_max_distance = 1.42
     couch_iso_size = 2.8
 
 
@@ -1308,6 +1318,7 @@ class Synthetic1mmOut1mmRight(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.41
     cax2bb_mean_distance = 1.3
     cax2bb_median_distance = 1.4
+    cax2epid_max_distance = 1.42
     couch_iso_size = 2.8
 
 
@@ -1317,6 +1328,7 @@ class Synthetic2mmUp1mmLeft(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 2.0
     cax2bb_mean_distance = 1.25
     cax2bb_median_distance = 1.0
+    cax2epid_max_distance = 2.02
     couch_iso_size = 2.0
 
 
@@ -1326,6 +1338,7 @@ class Synthetic2mmRight1mmDown(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 2.0
     cax2bb_mean_distance = 1.75
     cax2bb_median_distance = 2.0
+    cax2epid_max_distance = 2.02
     couch_iso_size = 4.0
 
 
@@ -1334,6 +1347,7 @@ class Synthetic1mmOut1SidedCouch(SyntheticWLMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_mean_distance = 1.0
     cax2bb_median_distance = 1
+    cax2epid_max_distance = 1.02
     couch_iso_size = 1.42
     images_axes = (
         (0, 0, 0),
@@ -1353,6 +1367,7 @@ class WLDemo(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.2
     cax2bb_median_distance = 0.7
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 2.36
     machine_scale = MachineScale.VARIAN_IEC
     epid_deviation = 1.3
     axis_of_rotation = {0: Axis.REFERENCE}
@@ -1447,8 +1462,9 @@ class WLLateral3mm(WinstonLutzMixin, TestCase):
     num_images = 4
     gantry_iso_size = 0.5
     cax2bb_max_distance = 3.8
-    cax2bb_median_distance = 2.3
+    cax2bb_median_distance = 2.24
     cax2bb_mean_distance = 2.3
+    cax2epid_max_distance = 3.68
     bb_shift_vector = Vector(x=-3.6, y=0.5, z=0.6)
 
 
@@ -1463,6 +1479,7 @@ class WLReferenceIsLargestRMS(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1
     cax2bb_median_distance = 0
     cax2bb_mean_distance = 0.33
+    cax2epid_max_distance = 1
     bb_shift_vector = Vector(x=1, y=0, z=0)
 
     def test_largest_error_at_ref_is_reported(self):
@@ -1477,8 +1494,9 @@ class WLLongitudinal3mm(WinstonLutzMixin, TestCase):
     num_images = 4
     gantry_iso_size = 0.5
     cax2bb_max_distance = 3.9
-    cax2bb_median_distance = 3.7
+    cax2bb_median_distance = 3.6
     cax2bb_mean_distance = 3.7
+    cax2epid_max_distance = 3.97
     bb_shift_vector = Vector(x=-0.63, y=3.6, z=0.6)
 
 
@@ -1486,9 +1504,10 @@ class WLVertical3mm(WinstonLutzMixin, TestCase):
     file_name = "vrt3mm.zip"
     num_images = 4
     gantry_iso_size = 0.5
-    cax2bb_max_distance = 3.8
+    cax2bb_max_distance = 3.7
     cax2bb_median_distance = 2.3
     cax2bb_mean_distance = 2.3
+    cax2epid_max_distance = 3.98
     bb_shift_vector = Vector(x=-0.5, y=0.5, z=3.6)
     print_results = True
 
@@ -1504,6 +1523,7 @@ class WLDontUseFileNames(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.9
     cax2bb_median_distance = 0.8
     cax2bb_mean_distance = 0.8
+    cax2epid_max_distance = 1.14
     bb_shift_vector = Vector(x=-0.4, y=0.6, z=0.6)
     axis_of_rotation = {
         0: Axis.REFERENCE,
@@ -1523,6 +1543,7 @@ class WLUseFileNames(WinstonLutzMixin, PlotlyTestMixin, TestCase):
     cax2bb_max_distance = 0.9
     cax2bb_median_distance = 0.8
     cax2bb_mean_distance = 0.8
+    cax2epid_max_distance = 1.14
     bb_shift_vector = Vector(y=0.6)
     axis_of_rotation = {
         0: Axis.COLLIMATOR,
@@ -1555,6 +1576,7 @@ class KatyiX0(WinstonLutzMixin, PlotlyTestMixin, TestCase):
     cax2bb_max_distance = 1.2
     cax2bb_median_distance = 0.8
     cax2bb_mean_distance = 0.7
+    cax2epid_max_distance = 1.3
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=-0.4, y=0.15, z=-0.5)
     print_results = True
@@ -1587,6 +1609,7 @@ class KatyiX1(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.2
     cax2bb_median_distance = 0.3
     cax2bb_mean_distance = 0.4
+    cax2epid_max_distance = 1.4
     bb_shift_vector = Vector(x=0.3, y=-0.2, z=0.3)
 
 
@@ -1599,6 +1622,7 @@ class KatyiX2(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.1
     cax2bb_median_distance = 0.5
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 1.1
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=0.15, y=-0.15, z=0.1)
 
@@ -1612,6 +1636,7 @@ class KatyiX3(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.25
     cax2bb_median_distance = 0.8
     cax2bb_mean_distance = 0.75
+    cax2epid_max_distance = 1.28
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=-0.1, y=0.2, z=-0.5)
 
@@ -1623,8 +1648,9 @@ class KatyTB0(WinstonLutzMixin, TestCase):
     collimator_iso_size = 0.8
     couch_iso_size = 1.3
     cax2bb_max_distance = 1.07
-    cax2bb_median_distance = 0.8
+    cax2bb_median_distance = 0.74
     cax2bb_mean_distance = 0.8
+    cax2epid_max_distance = 0.82
     machine_scale = MachineScale.VARIAN_IEC
     axis_of_rotation = {-1: Axis.REFERENCE}
     bb_shift_vector = Vector(x=-0.4, y=-0.1, z=-0.25)
@@ -1639,6 +1665,7 @@ class KatyTB1(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1
     cax2bb_median_distance = 0.7
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 0.76
     axis_of_rotation = {0: Axis.GANTRY}
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=-0.3, y=-0.2)
@@ -1653,6 +1680,7 @@ class KatyTB2(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.1
     cax2bb_median_distance = 0.4
     cax2bb_mean_distance = 0.5
+    cax2epid_max_distance = 0.5
     axis_of_rotation = {-1: Axis.REFERENCE}
     bb_shift_vector = Vector(x=0.0, y=-0.2, z=-0.6)
 
@@ -1667,6 +1695,7 @@ class ChicagoTBFinal(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.5
     cax2bb_median_distance = 0.3
     cax2bb_mean_distance = 0.3
+    cax2epid_max_distance = 0.26
     axis_of_rotation = {0: Axis.GBP_COMBO}
     bb_shift_vector = Vector(y=0.1)
 
@@ -1680,7 +1709,9 @@ class ChicagoTB52915(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.5
     cax2bb_median_distance = 0.3
     cax2bb_mean_distance = 0.3
+    cax2epid_max_distance = 0.31
     bb_shift_vector = Vector(z=0.2)
+    print_results = True
 
 
 class TrueBeam3120213(WinstonLutzMixin, TestCase):
@@ -1689,6 +1720,7 @@ class TrueBeam3120213(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.9
     cax2bb_median_distance = 0.35
     cax2bb_mean_distance = 0.4
+    cax2epid_max_distance = 0.81
     gantry_iso_size = 1.1
     collimator_iso_size = 0.7
     couch_iso_size = 0.7
@@ -1704,6 +1736,7 @@ class SugarLandiX2015(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.67
     cax2bb_median_distance = 1.05
     cax2bb_mean_distance = 1.1
+    cax2epid_max_distance = 1.66
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=0.6, y=-0.5, z=0.1)
 
@@ -1716,8 +1749,9 @@ class BayAreaiX0(WinstonLutzMixin, TestCase):
     collimator_iso_size = 1.1
     couch_iso_size = 2.3
     cax2bb_max_distance = 1.25
-    cax2bb_median_distance = 0.6
+    cax2bb_median_distance = 0.69
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 2.36
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=0, y=-0.3, z=-0.2)
 
@@ -1731,6 +1765,7 @@ class DAmoursElektaOffset(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 17.5
     cax2bb_median_distance = 14.3
     cax2bb_mean_distance = 13.8
+    cax2epid_max_distance = 16.64
     bb_shift_vector = Vector(x=10.2, y=-9.2, z=-11.1)  # independently verified
 
 
@@ -1743,6 +1778,7 @@ class DAmoursElektaXOffset(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 9.5
     cax2bb_median_distance = 6.9
     cax2bb_mean_distance = 6
+    cax2epid_max_distance = 9.93
     bb_shift_vector = Vector(x=-9.5, y=0.3, z=0.1)  # independently verified
 
 
@@ -1757,6 +1793,7 @@ class DAmoursElektaCentered(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.8
     cax2bb_median_distance = 0.5
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 2.23
     bb_shift_vector = Vector(y=0.4)
 
 
@@ -1771,6 +1808,7 @@ class DeBr6XElekta(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.0
     cax2bb_median_distance = 0.7
     cax2bb_mean_distance = 0.6
+    cax2epid_max_distance = 1.57
     bb_shift_vector = Vector(x=0.4, y=-0.2)
 
 
@@ -1785,6 +1823,7 @@ class LargeFieldCouchPresent(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 1.3
     cax2bb_median_distance = 1
     cax2bb_mean_distance = 1
+    cax2epid_max_distance = 1.11
     bb_shift_vector = Vector(x=0.5, y=-0.7, z=0.8)
 
 
@@ -1817,6 +1856,7 @@ class kVImages(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.26
     cax2bb_median_distance = 0.18
     cax2bb_mean_distance = 0.18
+    cax2epid_max_distance = 0.28
     axis_of_rotation = {-1: Axis.REFERENCE}
     bb_shift_vector = Vector(x=-0.24, y=0, z=0)
 
@@ -1840,6 +1880,7 @@ class TIFFImages(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.75
     cax2bb_median_distance = 0.5
     cax2bb_mean_distance = 0.5
+    cax2epid_max_distance = 0.67
     axis_of_rotation = {-1: Axis.GANTRY}
     bb_shift_vector = Vector(x=0.74, y=0.16, z=-0.14)
 
@@ -1858,6 +1899,7 @@ class VarianBBkV(WinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.42
     cax2bb_median_distance = 0.31
     cax2bb_mean_distance = 0.28
+    cax2epid_max_distance = 0.41
     axis_of_rotation = {-1: Axis.REFERENCE}
     bb_shift_vector = Vector(x=-0.24, y=0.2, z=-0.15)
 
@@ -1970,6 +2012,7 @@ class TestFrenchCBCT(CBCTWinstonLutzMixin, TestCase):
     cax2bb_max_distance = 0.23
     cax2bb_median_distance = 0.22
     cax2bb_mean_distance = 0.22
+    cax2epid_max_distance = 0.27
     machine_scale = MachineScale.VARIAN_IEC
     bb_shift_vector = Vector(x=0.2, y=0, z=-0.2)
 
@@ -1987,6 +2030,7 @@ class TestOffsetLeftCBCT(GeneratedWLCBCT, CBCTWinstonLutzMixin, TestCase):
     cax2bb_max_distance = 5
     cax2bb_median_distance = 2.5
     cax2bb_mean_distance = 2.5
+    cax2epid_max_distance = 5
     bb_shift_vector = Vector(x=-5, y=0, z=0)
 
 
@@ -1995,6 +2039,7 @@ class TestOffsetDownCBCT(GeneratedWLCBCT, CBCTWinstonLutzMixin, TestCase):
     cax2bb_max_distance = 5
     cax2bb_median_distance = 2.5
     cax2bb_mean_distance = 2.5
+    cax2epid_max_distance = 5
     bb_shift_vector = Vector(x=0, y=0, z=5)
 
 
@@ -2003,6 +2048,7 @@ class TestOffsetInCBCT(GeneratedWLCBCT, CBCTWinstonLutzMixin, TestCase):
     cax2bb_max_distance = 5
     cax2bb_median_distance = 5
     cax2bb_mean_distance = 5
+    cax2epid_max_distance = 5
     bb_shift_vector = Vector(x=0, y=-5, z=0)
 
 
@@ -2014,12 +2060,13 @@ class TestIndividualInverts(WinstonLutzMixin, TestCase):
     low_density_bb = True
     open_field = True
     bb_size = 5
-    gantry_iso_size = 0.57
+    gantry_iso_size = 0.32
     collimator_iso_size = None
     couch_iso_size = None
     cax2bb_max_distance = 0.55
     cax2bb_median_distance = 0.37
     cax2bb_mean_distance = 0.36
+    cax2epid_max_distance = 0.42
     axis_of_rotation = {-1: Axis.REFERENCE}
     bb_shift_vector = Vector(x=0.13, y=0.22, z=-0.3)
 
@@ -2036,3 +2083,75 @@ class TestIndividualInverts(WinstonLutzMixin, TestCase):
         for img in wl.images:
             img.crop(pixels=50)
         return wl
+
+
+class TestIsocal(TestCase):
+    """Test that if isocal is applied (3002,000D), the EPID position takes this into account."""
+
+    def test_isocal_synthetic(self):
+        shift = tempfile.TemporaryDirectory()
+        generate_winstonlutz(
+            simulator=AS1200Image(1000),
+            field_layer=PerfectFieldLayer,
+            dir_out=shift.name,
+            tags={"XRayImageReceptorTranslation": [1, 1, -500]},
+        )
+        no_shift = tempfile.TemporaryDirectory()
+        generate_winstonlutz(
+            simulator=AS1200Image(1000),
+            field_layer=PerfectFieldLayer,
+            dir_out=no_shift.name,
+            tags={"XRayImageReceptorTranslation": [0, 0, -500]},
+        )
+        wl_shift = WinstonLutz(shift.name)
+        wl_shift.analyze()
+        wl_shift_results = wl_shift.results_data()
+        # translation was set to 1mm in both x and y, thus sqrt(2) away
+        self.assertAlmostEqual(
+            wl_shift_results.max_2d_cax_to_epid_mm, 1.414, delta=0.01
+        )
+        wl_no_shift = WinstonLutz(no_shift.name)
+        wl_no_shift.analyze()
+        wl_no_shift_results = wl_no_shift.results_data()
+        self.assertAlmostEqual(wl_no_shift_results.max_2d_cax_to_epid_mm, 0)
+        # but that no other metric was changed
+        self.assertEqual(
+            wl_shift_results.max_gantry_rms_deviation_mm,
+            wl_no_shift_results.max_gantry_rms_deviation_mm,
+        )
+        self.assertEqual(
+            wl_shift_results.max_couch_rms_deviation_mm,
+            wl_no_shift_results.max_couch_rms_deviation_mm,
+        )
+        self.assertEqual(
+            wl_shift_results.max_coll_rms_deviation_mm,
+            wl_no_shift_results.max_coll_rms_deviation_mm,
+        )
+        self.assertEqual(
+            wl_shift_results.gantry_3d_iso_diameter_mm,
+            wl_no_shift_results.gantry_3d_iso_diameter_mm,
+        )
+
+
+class iXIsocal(WinstonLutzMixin, TestCase):
+    """An iX with a very large isocal offset"""
+
+    file_name = ["iX Linac with isocal.zip"]
+    bb_size = 10
+    num_images = 14
+    gantry_iso_size = 0.8
+    collimator_iso_size = None
+    couch_iso_size = None
+    cax2bb_max_distance = 0.73
+    cax2bb_median_distance = 0.33
+    cax2bb_mean_distance = 0.34
+    cax2epid_max_distance = 0.8
+
+
+class iXNoIsocal(iXIsocal):
+    """An iX without Isocal applied. Sister dataset to above made by stripping out 3002,000D tag.
+    The test inherits from iXIsocal, proving that all the values are the same except the cax2epid distance.
+    """
+
+    file_name = ["iX Linac isocal removed.zip"]
+    cax2epid_max_distance = 1.5  # only difference from the parent test